pyproject.toml
[tool.poetry]
name = "tiktok-business"
version = "0.1.0"
description = ""
authors = ["Chien Le <lchien063@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
google-cloud-bigquery = "^3.21.0"
google-cloud-tasks = "^2.16.3"
google-cloud-secret-manager = "^2.20.0"
uuid = "^1.30"
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
